Allow database to be kept when loading fixtures but still purged

+++ src/Services/DatabaseTools/ORMSqliteDatabaseTool.php
@@ -56,7 +56,7 @@ class ORMSqliteDatabaseTool extends ORMDatabaseTool
             }
         }
 
-        if (false === $append) {
+        if (false === $append && false === $this->getKeepDatabaseAndSchemaParameter()) {
             // TODO: handle case when using persistent connections. Fail loudly?
             $schemaTool = new SchemaTool($this->om);
             $schemaTool->dropDatabase();
